refactor: split scheduled learning contract
