+++ experiments/scheduled-task-learning/benchmark-core/benchmark_learning/learning_contract/__init__.py
@@ -1,0 +1,32 @@
+"""Closed ordered event and neutral adapter-receipt contracts for scheduled learning."""
+
+from .events import (
+    adapter_envelope_json,
+    canonical_event_log,
+    event_json,
+    parse_adapter_envelope,
+    parse_event,
+)
+from .receipts import decision_receipt, replay_receipt
+from .schema import (
+    AdapterEnvelope,
+    AdapterOutcome,
+    LearningContractError,
+    ReplayEvent,
+    ReplayEventKind,
+)
+
+__all__ = [
+    "AdapterEnvelope",
+    "AdapterOutcome",
+    "LearningContractError",
+    "ReplayEvent",
+    "ReplayEventKind",
+    "adapter_envelope_json",
+    "canonical_event_log",
+    "decision_receipt",
+    "event_json",
+    "parse_adapter_envelope",
+    "parse_event",
+    "replay_receipt",
+]
